requirements.txt
Django==4.2
django-filter==23.2
djangorestframework==3.14.0
psycopg2==2.9.6
schedule==1.2.0
APScheduler==3.10.1
django-scheduler~=0.10.1